test: update size golden for hello world lazy
The integration test is already close to the boundary. The actual
size in `main` is `236486`. We bump the threshold only by a few hundreds
(500b). Culprit is assumed to be the virtual instruction infrastructure
that is not three-shakable and impacts non-signal apps as well here.

it's only noticed here in this test as that one is already close to the
threshold.